Update index.html
I had all of this set up for tumblr, but since I couldn't get the video to work, I now need to (again) rewrite all the tab links so that instead of tumblr they are for github. I published the website on Tumblr, so that should work and have those links still saved for it specifically, it just doesn't have the video. RIP.