Update index.html
Rewrote the links on this home page. Now I need to trouble shoot. If it works, then do the same to the other pages manually (can't copy and paste unfortunately due to the class active code). If it doesn't work, *dramatic sigh*, then try and figure out what went wrong...